Fix suppression of Paginator continuation prompts.
This corrects an oversight - the continues/continued prompts were always
defined because the Paginator constructor defaulted them if they weren't
set in the parameters. Alas this meant the code to check if they should
be added couldn't be disabled by passing undef, and an empty string
would get turned into a blank line. Fix by changing the tests from
defined to truthy-value.