Improve handling of day names.
This dynamically calculates the name of a given day, rather than relying on a
fixed hash, and also implements a 'logical day' (akin to a 'broadcast
day') which rolls over at 3 a.m. (configurable) rather than midnight.
This allows events that start in the small hours to show up on the day
we expect them to, instead of the start of the real day's schedule.