Add support for generating 'sequential' page IDs.
The default behaviour of TelstarFrame's next_subpage method is to create
a new frame with the natural successor ID to the current one, so that
Telstar's default routing override for # kicks in.  However, because we
want to link back to specific sub-frames in places (e.g. from an event
to the schedule page it's listed on), and that requires unique page
numbers, add a 'sequential' mode that just increments the page number
for each new page and leaves the sub-frame ID at 'a'.

When in this mode, it also automatically wires up the # routing to the
new sub-frame, but this does require that the old frame is only saved
*after* creating the new frame. So update Paginator appropriately, too.