Reorganise schedule pages under the 100 range.
This both moves the schedules to the 100-series of numbers, and uses the
sequential-numbering option just introduced in TelstarFrame to give them
individual page numbers for each page of listing. Event pages now route
back to the correct listing sub-page and are found in the 11x range to
avoid collisions with the 10x listings range.

Also includes static redirect frames so *10x# points to the first page of
each listing (10x000 at the moment.)