Guestbook styling and script tweaks.
Set correct frame numbers, header and routing for guestbook pages. Also
allow the guestbook update script to run in both the primary and
secondary Telstar environments, updating the relevant server as
appropriate, as well as allow the guestbook to be regenerated without
adding a new message.

(TODO: separate guestbooks for each side?)