Minor layout tweak to price list pages.
This corrects issues where the manufacturer name occupies the entire
space available on the first line, by allowing wrapping on the colour
change and putting a better limit on the maximum indent for the
continuation lines.  Also widens the price field by a character since
some wine bottles break the £10 mark!