Add readback of event titles when adding reminder.
This adds generation of two audio clips for each event, one with just
the title and one with the full announcement including venue and
speaker list.  The filenames for these clips get stashed in the event
record in the database, from where it can be retrieved by the Lua code
and played to the user before the confirmation prompt.