Add outgoing calls when reminders fall due.
This consists of a `reminder_callback` function that needs to be hooked
up to an extension within `extensions.lua` and a despatcher script,
`depatch-reminders`, which needs to run periodically (ideally every
minute) and creates Asterisk call files that invoke said extension.

systemd unit and timer files suitable for this are included, although an
all-stars crontab entry would work just as well!