Dockerise EMF Live.
The EMF Live service is basically just a Telstar instance, but with a
response handler for the guestbook.  It'll be easier to maintain things
if we derive a new image from the Telstar one and bundle the guestbook
scripts into it, rather than the previous approach of copying them into
a volume by hand and trying to keep the Perl dependencies in sync.

This commit adds suitable Docker build bumf, and also tweaks the
guestbook frame and response handler to cope with their new locations.