Second draft of bar price list generation.
This required working, as the API endpoints that we were planning to hit
and pass into the first-draft scripts have been flagged as 'expensive,
please do not poll' - so we'd better not!

The 2024 bar API instead introduces a websocket-based subscription
mechanism, through which we can ask to be notified whenever a particular
item in the initial API response changes state.  That means we now run
as a long-running service instead of a one-shot script, and update the
pages as we receive notifications.

(Still to do: stare at the new API docs some more to make sure we're
handling the split between main and Null Sector bars properly, because
that seems to have changed and I'm convinced I've got this right yet!)