Correctly count separator markup in line length.
Telstar offers a series of markup sequences that produce a 39-character-
long line or dotted line; these are used with a lead-in mosaic-colour
control code to fill a complete line.  We need to account for this when
counting the line length because otherwise we'll put a carriage return
after it, introducing an unwanted blank line beneath.