Handle events where the header block doesn't fit.
There was only one event in the 2022 schedule that triggered this - by
having a very long 'max attendees' field - so rather than rework the
whole header section to be able to paginate it we can just truncate it
to fit.  Hopefully when I run this over the 2024 schedule tomorrow it
doesn't identify a whole slew of events with the same issue!